flask-user-auth-api/requirements.txt
flask
flask-cors
werkzeug
pyjwt
